Commit chieu thu 7 ngay 23.05.2014

+++ app/Controller/ChaptersController.php
@@ -53,7 +53,7 @@ class ChaptersController extends AppController {
         if (!$this->Chapter->exists($id)) {
             throw new NotFoundException(__('Invalid chapter'));
         }
-        $contain = array('TaiLieu','User' => array('fields' => array('id', 'name')), 'Field');
+        $contain = array('TaiLieu', 'User' => array('fields' => array('id', 'name')), 'Field');
         $options = array('conditions' => array('Chapter.' . $this->Chapter->primaryKey => $id), 'contain' => $contain);
         $this->set('chapter', $this->Chapter->find('first', $options));
     }
@@ -110,11 +110,11 @@ class ChaptersController extends AppController {
      * @return void
      */
     public function view($id = null) {
-        
+
         if (!$this->Chapter->exists($id)) {
             throw new NotFoundException(__('Invalid chapter'));
         }
-        $contain = array('TaiLieu','CreatedUser' => array('fields' => array('id', 'name')), 'Field');
+        $contain = array('TaiLieu', 'CreatedUser' => array('fields' => array('id', 'name')), 'Field');
         $options = array('conditions' => array('Chapter.' . $this->Chapter->primaryKey => $id), 'contain' => $contain);
         $this->set('chapter', $this->Chapter->find('first', $options));
     }
@@ -168,16 +168,42 @@ class ChaptersController extends AppController {
      * @return void
      */
     public function delete($id = null) {
+
         $this->Chapter->id = $id;
         if (!$this->Chapter->exists()) {
             throw new NotFoundException(__('Invalid chapter'));
         }
+
+
         $this->request->onlyAllow('post', 'delete');
-        if ($this->Chapter->delete()) {
-            return $this->flash(__('The chapter has been deleted.'), array('action' => 'index'));
+        if (!$this->Chapter->isOwnedBy($id, $this->Auth->user('id')) && (!$this->Chapter->User->isAdmin() || !$this->Chapter->User->isManager())) {
+            $this->Session->setFlash('Bạn không có quyền xóa chuyên đề người khác tạo', 'alert', array('plugin' => 'BoostCake', 'class' => 'alert-warning'));
+            return $this->redirect(array('action'=>'index'));
         } else {
-            return $this->flash(__('The chapter could not be deleted. Please, try again.'), array('action' => 'index'));
+            $course_number = $this->Chapter->field('course_number');
+            if ($course_number > 0) {
+                $this->Session->setFlash('Có ' . $course_number . ' khóa học thuộc chuyên đề này, bạn cần xóa chúng trước đã.', 'alert', array('plugin' => 'BoostCake', 'class' => 'alert-warning'));
+                return $this->redirect(array('action'=>'index'));
+            } else {
+                if ($this->Chapter->delete()) {
+                    $this->Session->setFlash('Xóa thành công', 'alert', array('plugin' => 'BoostCake', 'class' => 'alert-success'));
+                    return $this->redirect(array('action'=>'index'));
+                } else {
+                    $this->Session->setFlash('Xóa không thành công', 'alert', array('plugin' => 'BoostCake', 'class' => 'alert-warning'));
+                    return $this->redirect(array('action'=>'index'));
+                }
+            }
         }
     }
 
+    public function download($attachment_id){
+        $path=$this->Chapter->TaiLieu->getFilePath($attachment_id,'tai_lieu');
+        
+        $this->response->file(
+                $path, array('download' => true, 'name' => $this->Chapter->TaiLieu->getFileName($attachment_id))
+        );
+        // Return response object to prevent controller from trying to render
+        // a view
+        return $this->response;
+    }
 }
